Fix lots of warnings for MingW32. Only a few left. Compiles OK. Fix lots warnings for MingW64, but not yet compiling.

+++ src/ptclib/pstun.cxx
@@ -429,7 +429,7 @@ PSTUNAttribute * PSTUNMessage::GetFirstAttribute() const
   PSTUNAttribute * attr = (PSTUNAttribute *)(theArray+sizeof(PSTUNMessageHeader)); 
   PSTUNAttribute * ptr  = attr;
 
-  if ((CalcPaddedAttributeLength(attr->length) > GetSize()))
+  if ((CalcPaddedAttributeLength(attr->length) > (int)GetSize()))
     return NULL;
 
   while (ptr &&
